pyproject.toml
[project]
name = "rdms"
version = "0.1.0"
description = "E-Commerce Marketplace Database & Analytics"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
    "openai>=1.0.0",
    "pygraphviz>=1.11",
    "langgraph>=0.2.30",
    "langchain-core>=0.3.13",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langsmith>=0.1.0",
    "chromadb>=0.4.0",
    "pypdf>=3.0.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["script"]

